+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/sourcecontext/BundleSourcesTask.kt
@@ -59,12 +59,17 @@ abstract class BundleSourcesTask : Exec() {
     @get:Optional
     abstract val sentryProject: Property<String>
 
+    @get:Input
+    @get:Optional
+    abstract val sentryAuthToken: Property<String>
+
     override fun exec() {
         computeCommandLineArgs().let {
             commandLine(it)
             logger.info { "cli args: $it" }
         }
         setSentryPropertiesEnv()
+        setSentryAuthTokenEnv()
         super.exec()
     }
 
@@ -74,6 +79,15 @@ abstract class BundleSourcesTask : Exec() {
             environment("SENTRY_PROPERTIES", sentryProperties)
         } else {
             logger.info { "sentryProperties is null" }
+        }
+    }
+
+    internal fun setSentryAuthTokenEnv() {
+        val sentryAuthToken = sentryAuthToken.orNull
+        if (sentryAuthToken != null) {
+            environment("SENTRY_AUTH_TOKEN", sentryAuthToken)
+        } else {
+            logger.info { "sentryAuthToken is null" }
         }
     }
 
@@ -125,6 +139,7 @@ abstract class BundleSourcesTask : Exec() {
             cliExecutable: String,
             sentryOrg: String?,
             sentryProject: String?,
+            sentryAuthToken: Property<String>,
             taskSuffix: String = ""
         ): TaskProvider<BundleSourcesTask> {
             return project.tasks.register(
@@ -134,6 +149,7 @@ abstract class BundleSourcesTask : Exec() {
                 task.debug.set(debug)
                 task.sentryOrganization.set(sentryOrg)
                 task.sentryProject.set(sentryProject)
+                task.sentryAuthToken.set(sentryAuthToken)
                 task.sourceDir.set(collectSourcesTask.flatMap { it.output })
                 task.cliExecutable.set(cliExecutable)
                 SentryPropertiesFileProvider.getPropertiesFilePath(project, variant)?.let {