+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/BaseSentryPluginTest.kt
@@ -22,6 +22,7 @@ abstract class BaseSentryPluginTest(
     private lateinit var rootBuildFile: File
     protected lateinit var appBuildFile: File
     protected lateinit var moduleBuildFile: File
+    protected lateinit var sentryPropertiesFile: File
     protected lateinit var runner: GradleRunner
 
     protected open val additionalRootProjectConfig: String = ""
@@ -37,6 +38,7 @@ abstract class BaseSentryPluginTest(
 
         appBuildFile = File(testProjectDir.root, "app/build.gradle")
         moduleBuildFile = File(testProjectDir.root, "module/build.gradle")
+        sentryPropertiesFile = File(testProjectDir.root, "sentry.properties")
         rootBuildFile = testProjectDir.writeFile("build.gradle") {
             // language=Groovy
             """