+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Source Context for Java ([#495](https://github.com/getsentry/sentry-android-gradle-plugin/pull/495))
   - To enable it apply the `io.sentry.jvm.gradle` plugin and set `includeSourceContext` to `true`
   - For more information on how to enable source context, please refer to [#633](https://github.com/getsentry/sentry-java/issues/633#issuecomment-1465599120)
+- Allow setting sentry properties via the `sentry` plugin extension ([#500](https://github.com/getsentry/sentry-android-gradle-plugin/pull/500))
 
 ### Dependencies